requirements.txt
absl-py==2.2.2
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
argcomplete==3.6.2
astunparse==1.6.3
async-lru==2.0.5
attrs==24.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
boto3==1.38.16
botocore==1.38.16
Brotli==1.1.0
cachetools==5.5.2
cattrs==24.1.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.2
coremltools==8.2
CurrencyConverter==0.17.33
cycler==0.12.1
dacite==1.7.0
Deprecated==1.2.18
dill==0.4.0
dnspython==2.7.0
docx==0.2.4
dotenv==0.9.9
et_xmlfile==2.0.0
fastapi==0.115.4
fiftyone==1.5.2
fiftyone-brain==0.21.2
fiftyone_db==1.1.7
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
ftfy==6.3.1
future==1.0.0
gast==0.6.0
glob2==0.7
google-pasta==0.2.0
graphql-core==3.2.6
grpcio==1.73.1
h11==0.14.0
h2==4.2.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.6
httpx==0.27.2
hub-sdk==0.0.20
humanize==4.12.3
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
inflate64==1.0.1
jax==0.6.0
jaxlib==0.6.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
jsonlines==4.0.0
kaggle==1.7.4.5
kagglehub==0.3.12
kaleido==0.2.1
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
lxml==5.4.0
magic-filter==1.0.12
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
mongoengine==0.29.1
motor==3.6.1
mpmath==1.3.0
multidict==6.1.0
multivolumefile==0.2.3
namex==0.1.0
narwhals==1.39.1
networkx==3.4.2
numpy==2.1.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.2.3
pikepdf==9.7.0
pillow==11.2.1
plotly==6.1.0
pprintpp==0.4.0
priority==2.0.0
propcache==0.2.0
protobuf==4.25.7
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
py7zr==0.22.0
pyaml==25.1.0
pybcj==1.0.6
pycocotools==2.0.8
pycparser==2.22
pycryptodomex==3.22.0
pydantic==2.9.2
pydantic_core==2.23.4
pydash==8.0.5
Pygments==2.19.2
pymongo==4.9.2
pyparsing==3.2.3
pyppmd==1.1.1
pyTelegramBotAPI==4.13.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.1
python-slugify==8.0.4
python-telegram-bot==21.6
pytz==2025.2
PyYAML==6.0.2
pyzstd==0.17.0
rarfile==4.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==14.0.0
rtree==1.4.0
s3transfer==0.12.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soupsieve==2.7
SQLAlchemy==2.0.36
SQLite4==0.1.1
sse-starlette==0.10.3
sseclient-py==1.8.0
starlette==0.41.2
strawberry-graphql==0.268.1
sympy==1.13.3
tabulate==0.9.0
telebot==0.0.5
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
text-unidecode==1.3
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.147
ultralytics-thop==2.0.14
universal-analytics-python3==1.1.1
urllib3==2.2.3
uvicorn==0.32.0
voxel51-eta==0.14.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xmltodict==0.14.2
yarl==1.17.0
